+++ server/routes/users.js
@@ -1,19 +1,15 @@
 // users.js
 // Routes to CRUD users.
 
-var User = require('./user.js');
+var User = require('./_users.js');
 var path = require('path');
 var cookies = require('cookies');
-var keygrip = require('keygrip');
-var VCalendar = require('cozy-ical').VCalendar;
-var VEvent = require('cozy-ical').VEvent;
-var keys = keygrip(["Andres", "Franco"]);
-exports.CookKeys = keys;
 var fs = require('fs'); //FILESYSTEM
 var crypto = require('crypto');
 var nodemailer = require('nodemailer');
 var EmailTemplate = require('email-templates').EmailTemplate;
 var domain="https://127.0.1.1:3000";
+var secur = require('./secur.js');
 
 /******************************************************************************/
 /*                          LOAD TRANSLATIONS                                 */
@@ -30,133 +26,6 @@ fs.readFile(transFile, 'utf-8', function (err, t) {
         translations = JSON.parse(t);
     }
 });  
-
-/******************************************************************************/
-/*                          COOKIES & SECURITY                                */
-/******************************************************************************/
-
-/** 
- * Hash password
- * @param {String} pwd: The user's password
- * @param {String} salt: The user password's salt
- * @returns {Object} Password hash (.pass) and salt (.salt)
- */
-var hash = function (pwd, salt) {
-    if (!salt) {
-        try {
-            var buf = crypto.randomBytes(64);
-            salt = buf.toString('base64');
-        }
-        catch (ex) {
-            throw ex;
-        }
-    }
-    var pass = salt + pwd;
-
-    var passHash = crypto.createHash('sha256').update(pass).digest('base64');
-    var temp = {'pass': passHash, 'salt': salt};
-    return temp;
-};
-
-/**
- * Checks if the given ID matches the cookie ID. Call extractCookieData
- * before using this method!
- * @param {Integer} id: The user's ID
- * @param {Object} req: The HTTP request's headers
- * @param {Object} res: The HTTP request's response headers
- * @returns {Boolean} True if the IDs match false otherwise.
- */
-var sameUser = function (id, req, res) {
-
-    if (req.id == id) {
-        return true;
-    }
-    return false;
-};
-exports.sameUser = sameUser;
-
-/**
- * Checks the presence of an ID in the headers (if the user is logged in). Call 
- * extractCookieData before using this method!
- * @param {Object} req: The HTTP request's headers
- * @param {Object} res: The HTTP request's response headers
- * @returns {Boolean} True if the user is logged in, false otherwise.
- */
-var loggedIn = function (req, res) {
-
-    if (req.id) {
-        return true;
-    }
-    return false;
-};
-
-/**
- * Checks if the user doing the request is administrator.
- * @param {Integer} id: The user's ID
- * @param {Function} next: Function that executes next
- * @returns {callback} Runs next function.
- */
-var isAdmin = function (id, next) {
-
-    User.isAdmin(id, function (is) {
-        if (is)
-            return next(true);
-        else
-            return next(false);
-    });
-};
-exports.isAdmin = isAdmin;
-
-/**
-* Toma un string de formato 'XXhXX' (donde X es un dígito [0-9]) y devuelve
-* horas y minutos.
-* @param  {String} t
-* @return {Array} [hora,minutos]
-*/
-var getminutes = function (t) {
-   var mt = t.split('h');
-   return [parseInt(mt[0]), parseInt(mt[1])];
-};
-
-/**
-* Convierte un array de tipo [Year, Week, DayOfWeek] (ISO 8601) en un objeto de
-* tipo Date
-* @param {Array} AYWD ([Year, Week, DayOfWeek])
-* @return {Date Object} Dob
-*/
-var YWDarrToDob = function (AYWD) { // Arg : ISO 8601 : [Y, W, D]
-   var DOb = new Date(+AYWD[0], 0, 3);  // Jan 3
-   if (isNaN(DOb))
-       return false;
-   DOb.setDate(3 - DOb.getDay() + (AYWD[1] - 1) * 7 + +AYWD[2]);
-   return DOb;
-};
-
-/*
- * Method to extract cookies
- */
-/**
- * Extracts the cookies from the HTTP request header
- * @param {Object} req: The HTTP request's headers
- * @param {Object} res: The HTTP request's response headers
- * @param {Function} next: Function that executes next
- * @returns {callback} Runs next function.
- */
-exports.extractCookieData = function (req, res, next) {
-
-    var cook = new cookies(req, res, keys);
-    var idCookie = cook.get('LinkedEnibId');
-
-    if (idCookie) {
-
-        req.id = parseInt(idCookie);
-    }
-    else {
-        console.log('Cookies Errors');
-    }
-
-    return next();
-};
 
 /**
  * Sends mail.
@@ -222,10 +91,6 @@ var sendActivationEmail = function(email,hash,name,lastname){
 /*                          GET METHODS                                       */
 /******************************************************************************/
 
-exports.getPlaces = function(req, res, next){
-  res.status(501).send('Not Implemented');
-};
-
 exports.getTranslation = function(req,res,next){
     
     var lang = req.params.lang;
@@ -237,150 +102,6 @@ exports.getTranslation = function(req,res,next){
         res.status(404).send('translation for:'+lang+'not found');
     }
 
-};
-
-exports.getEdtConfig = function(req, res, next){
-    
-    //TODO: Configuracion personal del usuario
-    
-    var configFile = path.resolve(path.join(__dirname, '../config/edtConfig.json'));
-    
-    fs.readFile(configFile, 'utf-8', function (err, config) {
-        if (err){
-            console.log(err);
-            res.sendStatus(500);
-        }
-        else{
-            console.log(config);
-            res.status(200).send({config: JSON.parse(config)});
-        }
-    });  
-};
-
-exports.getTimesIcal = function(req, res, next){
-    
-    if( req.query.whatId && req.query.whoId && 
-        req.query.week && req.query.year && req.query.whatName)
-        ;
-    else{
-        res.status(401).send('Missing information');
-        return;
-    }
-    
-    var timeData = {
-        //TODO: Extender a array de whats
-        whatName: req.query.whatName,
-        //whoName: req.query.whoName,
-        whatId: req.query.whatId,
-        whoId: req.query.whoId,
-        week: req.query.week,
-        year: req.query.year,
-    };
-      
-    User.getTimes(timeData, function (err, times) {
-        if (err) {
-            console.log(err);
-            res.status(500).send('Error');
-            return;
-        } else {
-            console.log(times);
-            var cal = new VCalendar({
-                organization: 'TimesApp',
-                title: timeData.whatName
-            });
-            var tmpDate,tmpHMin,startDate,endDate;
-            times.forEach(function(time){
-                tmpDate = YWDarrToDob([time.year,time.week,time.day])
-                tmpHMin = getminutes(time.from);
-                startDate = new Date(   tmpDate.getFullYear(), 
-                                        tmpDate.getMonth(),
-                                        tmpDate.getDate(),
-                                        tmpHMin[0]+Math.floor(time.timezone/60), tmpHMin[1]+time.timezone%60, 0);
-                tmpHMin = getminutes(time.to);                        
-                endDate = new Date(     tmpDate.getFullYear(), 
-                                        tmpDate.getMonth(),
-                                        tmpDate.getDate(),
-                                        tmpHMin[0]+Math.floor(time.timezone/60), tmpHMin[1]+time.timezone%60, 0);
-                                        
-                var vevent = new VEvent({
-                    stampDate: 0,
-                    startDate: startDate,
-                    endDate: endDate,
-                    summary: time.type,
-                    description: time.desc,
-                    //location: "Test Location",
-                    uid: time.idNEO
-                });
-
-                cal.add(vevent);
-            });
-
-            console.log(cal.toString());
-
-            fs.writeFile("/tmp/ical.ics", cal.toString(), function(err) {
-                if(err) {
-                    return console.log(err);
-                    res.sendStatus(500);
-                }else{
-                    var stat = fs.statSync('/tmp/ical.ics');
-                    res.writeHead(200, {
-                          'Content-Type': 'text/calendar', 
-                          'Content-Length': stat.size,
-                          'Content-disposition' : 'attachment; filename="calendar.ics"'
-                    });
-                    var stream = fs.createReadStream( '/tmp/ical.ics', { bufferSize: 64 * 1024 });
-                    //res.attachment('/tmp/ical.ics');
-                    stream.pipe(res);
-        //            res.download('/tmp/ical.ics','calendar.ics');
-                }
-            }); 
-        }
-    });
-};
-
-exports.getTimes = function(req, res, next){
-  
-    var timeData = {
-        whatId: req.query.whatId,
-        whoId: req.query.whoId,
-        week: req.query.week,
-        year: req.query.year,
-    };
-      
-    User.getTimes(timeData, function (err, times) {
-        if (err) {
-            console.log(err);
-            res.status(500).send('Error');
-            return;
-        } else {
-            console.log(times);
-            res.status(200).send({times: times});
-            return;
-        }
-    });
-};
-
-exports.getActivityTypes = function (req, res, next) {
-
-  if (!req.id) {
-    res.status(401).send('Unauthorized');
-    return;
-  } else if (!req.query.parent) {
-    res.status(400).send('Missing parent');
-    return;
-  }
-
-  User.getActivityTypes(req.query.parent, function (err, activityTypes) {
-    if (err) {
-      console.log(err);
-      res.status(500).send('Error');
-      return;
-    } else {
-      console.log(activityTypes);
-      res.status(200).send({activityTypes: activityTypes});
-      return;
-    }
-  });
 };
 
 exports.getAsocs = function (req, res, next) {
@@ -418,7 +139,7 @@ exports.getProfile = function (req, res, next) {
 
     if (req.params.id) {
         idNEO = parseInt(req.params.id);
-        public = !sameUser(idNEO, req, res);
+        public = !secur.sameUser(idNEO, req, res);
     } else if (req.id) {
         idNEO = req.id;
         public = false;
@@ -513,78 +234,6 @@ exports.getNodeContents = function (req, res, next) {
         return;
     });
 
-};
-
-/**
- * Sends the ADMIN's parts.
- * @param {Object} req: The HTTP request's headers
- * @param {Object} res: The HTTP request's response headers
- * @param {Function} next: Function that executes next
- * @returns {void} Nothing.
- */
-exports.getAdminNodes = function (req, res, next) {
-
-    if (req.id) {
-        ;
-    } else {
-        res.status(401).send('Unauthorized');
-        return;
-    }
-
-    User.getAdminNodes(req.id, function (err, adminnodes) {
-        if (err) {
-            res.status(500).send('Error');
-            return;
-        }
-        if (adminnodes) {
-            res.status(200).send({adminnodes: adminnodes});
-            return;
-        }
-        res.status(500).send('Error');
-        return;
-    });
-};
-
-exports.getNodeRelTypes = function(req, res, next){
-
-  if(!req.id){
-    res.status(401).send('Unauthorized');
-    return;
-  } else if(!req.query.memberof || req.query.memberof == ''){
-    res.status(400).send('Missing MemberOf');
-    return;
-  }
-
-  User.getNodeRelTypes(req.query.memberof, function(err, nodeTypes, relTypes){
-    if(err || !nodeTypes || !relTypes){
-      res.status(500).send('Error getting Node Rel Types');
-      return;
-    } else{
-      res.status(200).send({nodetypes: nodeTypes, reltypes: relTypes});
-    }
-  });
-};
-
-exports.getNodeRelFields = function(req, res, next){
-
-  if(!req.id){
-    res.status(401).send('Unauthorized');
-    return;
-  } else if(!req.params.label){
-    res.status(400).send('Missing Label');
-    return;
-  }
-
-  User.getNodeRelFields(req.params.label, function(err, fields){
-    if(err){
-      res.status(500).send('Error getting fields');
-      return;
-    }else{
-      console.log(fields);
-      res.status(200).send({fields: fields});
-      return;
-    }
-  });
 };
 
 /**
@@ -790,23 +439,9 @@ exports.subscribe = function(req, res, next){
     });
 };
 
-exports.newActivity = function (req, res, next) {
-    var acts = req.body.activities;
-
-    User.newActivity(acts, function (err, result) {
-        if (err) {
-            res.status(500).send('Error Creating Activities');
-            return;
-        } else {
-            res.status(200).send('OK');
-            return;
-        }
-    });
-};
-
 exports.updateProfile = function (req, res, next) {
 
-    if (!sameUser(req.body.id, req, res)) {
+    if (!secur.sameUser(req.body.id, req, res)) {
         res.status(401).send('Unauthorized');
         return;
     }
@@ -820,70 +455,6 @@ exports.updateProfile = function (req, res, next) {
             return;
         } else {
             res.status(200).send(profile);
-            return;
-        }
-    });
-};
-
-/**
- * POST /newrel
- */
-exports.newRel = function (req, res, next) {
-    var relData = req.body;
-    if (!relData.hasOwnProperty('instId')) {
-        res.status(400).send('Missing Organism');
-        return;
-    }
-
-    if (!relData.hasOwnProperty('usrID')) {
-        res.status(400).send('Missing User');
-        return;
-    }
-
-    if (!relData.hasOwnProperty('relType')) {
-        res.status(400).send('Missing Relationship Details');
-        return;
-    }
-    console.log(relData);
-    User.newRel(relData, function (err) {
-        if (err) {
-            res.status(400).send('Error Creating Rel');
-            return;
-        } else {
-            res.status(200).send('OK');
-        }
-    });
-};
-
-/**
- * POST /newpart
- */
-exports.newPart = function (req, res, next) {
-    var data = req.body;
-    if (!data.hasOwnProperty('instID')) {
-        res.status(400).send('Missing Organism');
-        return;
-    }
-
-    if (!data.hasOwnProperty('partData')) {
-        res.status(400).send('Missing Node Data');
-        return;
-    }
-
-    if (!data.hasOwnProperty('label')) {
-        res.status(400).send('Missing Node Label');
-        return;
-    }
-
-    User.newPart(data, function (err, partID) {
-        if (err) {
-            res.status(400).send('Error Creating Node');
-            return;
-        } else if (partID) {
-            res.status(200).send({idNEO: partID});
-            return;
-        } else {
-            res.status(500).send('Database error');
             return;
         }
     });
@@ -905,7 +476,7 @@ exports.signup = function (req, res, next) {
     }
     
     var email = nodeData.email;
-    var tempPass = hash(nodeData['password'], null);
+    var tempPass = secur.hash(nodeData['password'], null);
     nodeData.password = tempPass['pass'];
     nodeData.salt = tempPass['salt'];
     //nodeData['active'] = 0;
@@ -958,7 +529,7 @@ exports.login = function (req, res, next) {
             return;
         }
 
-        var secPass = hash(nodeData['password'], results['salt']);
+        var secPass = secur.hash(nodeData['password'], results['salt']);
 
         if (secPass['pass'] !== results['pass']) {
             res.status(401).send('Wrong email or password');
@@ -966,8 +537,8 @@ exports.login = function (req, res, next) {
         }
         ;
         if (results['idNEO']) {
-            if (!loggedIn(req, res)) {
-                var cook = new cookies(req, res, keys);
+            if (!secur.loggedIn(req, res)) {
+                var cook = new cookies(req, res, secur.cookKeys);
                 cook.set('LinkedEnibId', results.idNEO, {signed: true, maxAge: 9000000});
                 console.log(results['idNEO']);
                 res.status(200).send({idNEO: results['idNEO'], lang: results.lang});
@@ -1010,7 +581,7 @@ exports.friend = function (req, res, next) {
  */
 exports.uploadPic = function (req, res, next) {
     var id = req.params.id;
-    if (!sameUser(id, req, res)) {
+    if (!secur.sameUser(id, req, res)) {
         res.status(401).send('Unauthorized');
         return;
     }
@@ -1076,28 +647,9 @@ exports.changeProperty = function (req, res, next) {
     });
 };
 
-exports.verifyPassword = function (req, res, next) {
-    
-    User.verifyPassword(req.body['id'], function (err, results) {
-        if (err) {
-            res.status(500).send('Error');
-            return;
-        }else{
-            var tmp = hash(req.body['password'], results['salt']);
-            if (tmp['pass'] !== results['pass']) {
-                res.status(401).send('Wrong username or password');
-                return;
-            }else{
-                next();
-                return;
-            }
-        }
-    });
-};
-
 exports.changePassword = function (req, res, next) {
 
-    var pswdNew = hash(req.body['new'], null);
+    var pswdNew = secur.hash(req.body['new'], null);
 
     User.changePassword(pswdNew['pass'], pswdNew['salt'], req.body['id'], function (err) {
         if (err) {
@@ -1140,7 +692,7 @@ exports.deleteFriend = function (req, res, next) {
  * DELETE /delUser/:id -> DELETE USER
  */
 exports.deleteUser = function (req, res, next) {
-    if (!sameUser(req.params.id, req, res)) {
+    if (!secur.sameUser(req.params.id, req, res)) {
         res.status(401).send('Unauthorized');
         return;
     }
